packages/ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
xgboost==2.0.3
pydantic==2.5.3
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==6.1.0
httpx==0.26.0
psycopg2-binary==2.9.9
